+++ pkg/apis/deployment/v1/plan.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2021 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,8 +16,6 @@
 // limitations under the License.
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
-//
-// Author Ewout Prangsma
 //
 
 package v1
@@ -299,6 +297,11 @@ func NewActionBuilder(group ServerGroup, memberID string) ActionBuilder {
 func (a Action) SetImage(image string) Action {
 	a.Image = image
 	return a
+}
+
+// IsStarted returns true if the action has been started already.
+func (a Action) IsStarted() bool {
+	return !a.StartTime.IsZero()
 }
 
 // AsPlan parse action list into plan