+++ pkg/apis/deployment/v2alpha1/register.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2021 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,8 +16,6 @@
 // limitations under the License.
 //
 // Copyright holder is ArangoDB GmbH, Cologne, Germany
-//
-// Author Ewout Prangsma
 //
 
 package v2alpha1
@@ -52,6 +50,8 @@ func addKnownTypes(s *runtime.Scheme) error {
 		&ArangoDeploymentList{},
 		&ArangoMember{},
 		&ArangoMemberList{},
+		&ArangoClusterSynchronization{},
+		&ArangoClusterSynchronizationList{},
 	)
 	metav1.AddToGroupVersion(s, SchemeGroupVersion)
 	return nil